+++ janitor/lib/publisher.py
@@ -2,10 +2,11 @@ from pyxavi.config import Config
 from pyxavi.logger import Logger
 from pyxavi.terminal_color import TerminalColor
 from pyxavi.mastodon_publisher import MastodonPublisher, MastodonPublisherException
+from pyxavi.queue_stack import Queue
 from janitor.objects.queue_item import QueueItem
 from janitor.objects.message import Message, MessageType
-from .queue import Queue
 from .formatter import Formatter
+import os
 
 
 class Publisher(MastodonPublisher):
@@ -17,6 +18,7 @@ class Publisher(MastodonPublisher):
 
     MAX_RETRIES = 3
     SLEEP_TIME = 10
+    DEFAULT_QUEUE_FILE = "storage/queue.yaml"
 
     def __init__(
         self,
@@ -26,14 +28,18 @@ class Publisher(MastodonPublisher):
         only_oldest: bool = None
     ) -> None:
 
+        logger = Logger(config=config).get_logger()
+
         super().__init__(
-            config=config,
-            logger=Logger(config=config).get_logger(),
-            named_account=named_account,
-            base_path=base_path
+            config=config, logger=logger, named_account=named_account, base_path=base_path
         )
 
-        self._queue = Queue(config, base_path=base_path)
+        queue_storage_file = config.get("queue_storage.file", self.DEFAULT_QUEUE_FILE)
+        if base_path is not None:
+            queue_storage_file = os.path.join(base_path, queue_storage_file)
+        self._queue = Queue(
+            logger=logger, storage_file=queue_storage_file, queue_item_object=QueueItem
+        )
         self._formatter = Formatter(config, self._connection_params.status_params)
         # Janitor has the dry_run set up somewhere else. Overwriting.
         self._is_dry_run = config.get("app.run_control.dry_run", False)